# docs/flow-fic-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FlowFic API",
    "version": "2026-03-16",
    "description": "OpenAPI spec generated from route handlers in src/app/api and supabase/functions for import into Apidog."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Next.js app"
    },
    {
      "url": "https://your-domain.com",
      "description": "Production Next.js app"
    }
  ],
...
  "paths": {
...
    "/api/admin/payments/topup-coins": {
      "post": {
        "tags": [
          "Admin Payments"
        ],
        "operationId": "adminTopupCoins",
        "summary": "Top up coins to a selected user",
        "security": [
          {
            "SupabaseBearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminTopupCoinsRequest"
              },
              "example": {
                "userId": "1abcc2d6-92b4-4b34-b21c-33fb8eda92ec",
                "amount": 100,
                "reason": "Compensate outage incident",
                "correlationId": "topup-20260328-001"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Top-up applied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTopupCoinsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      }
    },
...
    "/api/admin/payments/apply-chargeback-hold": {
      "post": {
        "tags": [
          "Admin Payments"
        ],
        "operationId": "applyChargebackHold",
        "summary": "Apply a chargeback hold to a user's coin balance",
        "security": [
          {
            "SupabaseBearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyChargebackHoldRequest"
              },
              "example": {
                "userId": "user_123",
                "amount": 120,
                "reason": "Stripe chargeback case opened",
                "externalReference": "chb_001",
                "correlationId": "chargeback-case-001"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hold applied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyChargebackHoldResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      }
    },
...
    "/api/admin/payouts/approve": {
      "post": {
        "tags": [
          "Admin Payouts"
        ],
        "summary": "Approve creator payout request",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/admin/payouts/reject": {
      "post": {
        "tags": [
          "Admin Payouts"
        ],
        "summary": "Reject creator payout request",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/admin/payouts/mark-paid": {
      "post": {
        "tags": [
          "Admin Payouts"
        ],
        "summary": "Mark creator payout request as paid",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "SupabaseBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase access token in the Authorization header."
      },
      "CronBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token",
        "description": "CRON_SECRET in the Authorization header for internal cron-protected finance endpoints."
      },
      "StripeSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "stripe-signature",
        "description": "Stripe webhook signature over the raw request body."
      }
    },
    "responses": {
      "Error400": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Error401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Error403": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Error404": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Error409": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Error500": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "allowed": {
            "type": "boolean"
          },
          "stripeStatus": {
            "type": "integer"
          },
          "currentBalance": {
            "type": "number"
          },
          "requiredBalance": {
            "type": "number"
          },
          "requestId": {
            "type": "string"
          },
          "duplicate": {
            "type": "boolean"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": true
      },
...
      "AdminTopupCoinsRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000000
          },
          "reason": {
            "type": "string",
            "minLength": 8
          },
          "correlationId": {
            "type": "string",
            "maxLength": 120
          }
        },
        "required": [
          "userId",
          "amount",
          "reason"
        ],
        "additionalProperties": false
      },
      "AdminTopupCoinsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "integer"
          },
          "transactionId": {
            "type": "string",
            "nullable": true
          },
          "newBalance": {
            "type": "number"
          },
          "correlationId": {
            "type": "string"
          },
          "policyVersion": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "userId",
          "amount",
          "transactionId",
          "newBalance",
          "correlationId",
          "policyVersion"
        ],
        "additionalProperties": false
      },
...
      "ApplyChargebackHoldRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0
          },
          "reason": {
            "type": "string",
            "minLength": 8
          },
          "externalReference": {
            "type": "string"
          },
          "correlationId": {
            "type": "string",
            "maxLength": 120
          }
        },
        "required": [
          "userId",
          "amount",
          "reason"
        ],
        "additionalProperties": false
      },
      "ApplyChargebackHoldResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "paymentCaseId": {
            "type": "string"
          },
          "holdTransactionId": {
            "type": "string",
            "nullable": true
          },
          "newBalance": {
            "type": "number"
          },
          "financeStatus": {
            "type": "string",
            "enum": [
              "restricted_finance"
            ]
          },
          "policyVersion": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "paymentCaseId",
          "holdTransactionId",
          "newBalance",
          "financeStatus",
          "policyVersion"
        ]
      },
...
    }
  }
}